move functional to preprocessing_state.py
